requirements.txt
absl-py==1.4.0
anyio==3.6.2
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
cryptography==40.0.2
dlib==19.24.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.95.1
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.5.0
keras==2.11.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.9
starlette==0.26.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
#tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
